+++ webapp/app.py
@@ -2,12 +2,25 @@
 A Flask application for maas.io
 """
 
+# Standard library
 import flask
 
+# Packages
 from canonicalwebteam.flask_base.app import FlaskBase
 from canonicalwebteam.templatefinder import TemplateFinder
 
-from webapp.feeds import get_rss_feed
+# Local
+from webapp.context import (
+    build_path_with_params,
+    current_year,
+    descending_years,
+    format_date,
+    get_json_feed,
+    month_name,
+    months_list,
+    navigation,
+    releases,
+)
 
 
 app = FlaskBase(
@@ -34,4 +47,14 @@ def internal_error(error):
 
 @app.context_processor
 def context():
-    return dict(get_rss_feed=get_rss_feed)
+    return {
+        "build_path_with_params": build_path_with_params,
+        "current_year": current_year,
+        "descending_years": descending_years,
+        "format_date": format_date,
+        "get_json_feed": get_json_feed,
+        "month_name": month_name,
+        "months_list": months_list,
+        "navigation": navigation,
+        "releases": releases(),
+    }
